requirements.txt
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
torch>=2.3.0
torchvision>=0.18.0
pillow>=10.4.0
numpy>=1.26.0
scikit-image>=0.24.0
matplotlib>=3.9.0
pyyaml>=6.0.2
